+++ examples/showoci/showoci.py
@@ -118,7 +118,7 @@ import contextlib
 import os
 import time
 
-version = "23.09.03"
+version = "23.09.26"
 
 ##########################################################################
 # check OCI version
@@ -349,6 +349,7 @@ def set_parser_arguments(argsList=[]):
     parser.add_argument('-mc', action='store_true', default=False, dest='mgdcompart', help='exclude ManagedCompartmentForPaaS')
     parser.add_argument('-nr', action='store_true', default=False, dest='noroot', help='Not include root compartment')
     parser.add_argument('-ip', action='store_true', default=False, dest='instance_principals', help='Use Instance Principals for Authentication')
+    parser.add_argument('-rp', action='store_true', default=False, dest='resource_principals', help='Use Resource Principals for Authentication')
     parser.add_argument('-is', action='store_true', default=False, dest='security_token', help='Use Config and Security Token for Authentication')
     parser.add_argument('-dt', action='store_true', default=False, dest='delegation_token', help='Use Delegation Token (Cloud shell)')
     parser.add_argument('-t', default="", dest='profile', help='Config file section to use (tenancy profile)')
@@ -505,6 +506,9 @@ def set_service_extract_flags(cmd):
 
     if cmd.instance_principals:
         prm.use_instance_principals = True
+
+    if cmd.resource_principals:
+        prm.use_resource_principals = True
 
     if cmd.delegation_token:
         prm.use_delegation_token = True