+++ src/oci/data_science/models/notebook_session_configuration_details.py
@@ -32,6 +32,10 @@ class NotebookSessionConfigurationDetails(object):
             The value to assign to the subnet_id property of this NotebookSessionConfigurationDetails.
         :type subnet_id: str
 
+        :param private_endpoint_id:
+            The value to assign to the private_endpoint_id property of this NotebookSessionConfigurationDetails.
+        :type private_endpoint_id: str
+
         :param notebook_session_shape_config_details:
             The value to assign to the notebook_session_shape_config_details property of this NotebookSessionConfigurationDetails.
         :type notebook_session_shape_config_details: oci.data_science.models.NotebookSessionShapeConfigDetails
@@ -41,6 +45,7 @@ class NotebookSessionConfigurationDetails(object):
             'shape': 'str',
             'block_storage_size_in_gbs': 'int',
             'subnet_id': 'str',
+            'private_endpoint_id': 'str',
             'notebook_session_shape_config_details': 'NotebookSessionShapeConfigDetails'
         }
 
@@ -48,12 +53,14 @@ class NotebookSessionConfigurationDetails(object):
             'shape': 'shape',
             'block_storage_size_in_gbs': 'blockStorageSizeInGBs',
             'subnet_id': 'subnetId',
+            'private_endpoint_id': 'privateEndpointId',
             'notebook_session_shape_config_details': 'notebookSessionShapeConfigDetails'
         }
 
         self._shape = None
         self._block_storage_size_in_gbs = None
         self._subnet_id = None
+        self._private_endpoint_id = None
         self._notebook_session_shape_config_details = None
 
     @property
@@ -131,6 +138,30 @@ class NotebookSessionConfigurationDetails(object):
         :type: str
         """
         self._subnet_id = subnet_id
+
+    @property
+    def private_endpoint_id(self):
+        """
+        Gets the private_endpoint_id of this NotebookSessionConfigurationDetails.
+        The OCID of a Data Science private endpoint.
+
+
+        :return: The private_endpoint_id of this NotebookSessionConfigurationDetails.
+        :rtype: str
+        """
+        return self._private_endpoint_id
+
+    @private_endpoint_id.setter
+    def private_endpoint_id(self, private_endpoint_id):
+        """
+        Sets the private_endpoint_id of this NotebookSessionConfigurationDetails.
+        The OCID of a Data Science private endpoint.
+
+
+        :param private_endpoint_id: The private_endpoint_id of this NotebookSessionConfigurationDetails.
+        :type: str
+        """
+        self._private_endpoint_id = private_endpoint_id
 
     @property
     def notebook_session_shape_config_details(self):